requirements.txt
customtkinter~=5.2.2
CTkMessagebox~=2.5
bs4~=0.0.2
beautifulsoup4~=4.12.3
langchain~=0.1.16
openai~=1.17.1
python-dotenv~=1.0.1